+++ lib/td_dq_web/controllers/rule_result_controller.ex
@@ -97,7 +97,7 @@ defmodule TdDqWeb.RuleResultController do
       rule_implementation ->
         case get_concept(rule_implementation) do
           nil ->
-            []
+            ""
 
           concept ->
             concept